pyproject.toml
[tool.poetry]
name = "ai-build-hackathon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.2.5"
langchain = "^0.2.3"
langchain-cohere = "^0.1.5"
python-dotenv = "^1.0.1"
langgraph = "^0.0.65"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
